+++ src/local-development/helpers/md5.ts
@@ -1,0 +1,5 @@
+import crypto from 'node:crypto';
+
+export function md5(input: string) {
+	return crypto.createHash('md5').update(input).digest('hex');
+}