+++ src/local-development/pull.ts
@@ -11,7 +11,7 @@ import {
 	updateMakecomappJson,
 	upsertComponentInMakecomappjson,
 } from './makecomappjson';
-import { convertComponentMetadataRemoteNamesToLocalIds, getRemoteComponentsSummary } from './remote-components-summary';
+import { convertComponentMetadataRemoteNamesToLocalIds, getRemoteComponent } from './remote-components-summary';
 import { generateComponentDefaultCodeFilesPaths } from './local-file-paths';
 import { pullComponentCode, pullComponentCodes } from './code-pull-deploy';
 import { askForProjectOrigin } from './ask-origin';
@@ -23,6 +23,8 @@ import { AppComponentType } from '../types/app-component-type.types';
 import { catchError } from '../error-handling';
 import { withProgressDialog } from '../utils/vscode-progress-dialog';
 import { entries } from '../utils/typed-object';
+import { Checksum } from './types/checksum.types';
+import { downloadOriginChecksums, getComponentChecksumArray } from './helpers/origin-checksum';
 
 export function registerCommands(): void {
 	/**
@@ -37,7 +39,8 @@ export function registerCommands(): void {
 				return;
 			}
 			await withProgressDialog({ title: '' }, async (_progress, _cancellationToken) => {
-				await pullAllComponents(localAppRootdir, origin, 'askUser');
+				const originChecksums = await downloadOriginChecksums(origin);
+				await pullAllComponents(localAppRootdir, origin, 'askUser', originChecksums);
 			});
 			vscode.window.showInformationMessage('All local app component files has been updated.');
 		}),
@@ -53,6 +56,7 @@ export async function pullAllComponents(
 	localAppRootdir: vscode.Uri,
 	origin: LocalAppOriginWithSecret,
 	newRemoteComponentResolution: 'askUser' | 'cloneAsNew',
+	originChecksums: Checksum,
 ): Promise<void> {
 
 	// Pulling when there are undeployed changes. Override them and align the state with the origin.
@@ -61,15 +65,8 @@ export async function pullAllComponents(
 	componentIdMappingHelper.filterMappingItems(item => !item?.localDeleted);
 	await updateMakecomappJson(localAppRootdir, makecomappJson);
 
-	const remoteAppComponentsSummary = await getRemoteComponentsSummary(localAppRootdir, origin);
 	// Compare all local components with remote. If something is not paired, link it or create new component or ignore component.
-	await alignComponentsMapping(
-		localAppRootdir,
-		origin,
-		remoteAppComponentsSummary,
-		'ignore',
-		newRemoteComponentResolution,
-	);
+	await alignComponentsMapping(localAppRootdir, origin, originChecksums, 'ignore', newRemoteComponentResolution);
 	// Load fresh `makecomapp.json` file (because `alignComponentMapping()` changed it)
 	const makecomappJsonFile = await MakecomappJsonFile.fromLocalProject(localAppRootdir);
 
@@ -95,9 +92,12 @@ export async function pullAllComponents(
 	// Note: All remote components already have local counterparts due to previously called `alignComponentsMapping()`.
 	for (const componentType of getAppComponentTypes()) {
 		const componentIdMapping = new ComponentIdMappingHelper(makecomappJsonFile.content, origin);
-		for (const [remoteComponentName, remoteComponentMetadata] of Object.entries(
-			remoteAppComponentsSummary[componentType],
-		)) {
+		const checksums = getComponentChecksumArray(originChecksums, componentType);
+		const originNames = checksums.map((checksum) => {
+			return checksum.name;
+		});
+
+		for (const remoteComponentName of originNames) {
 			const componentLocalId = componentIdMapping.getExistingLocalId(componentType, remoteComponentName);
 			if (componentLocalId === null) {
 				continue;
@@ -112,13 +112,14 @@ export async function pullAllComponents(
 			}
 
 			// Prepate updated component metadata (merge with previous one)
+			const remoteComponentMetadata = await getRemoteComponent(origin, componentType, remoteComponentName);
 			const updatedComponentMedatada: AppComponentMetadataWithCodeFiles = {
 				...existingComponentMetadata, // Use previous `codeFiles`
 				// + Update all other properties by fresh one loaded from remote
 				...convertComponentMetadataRemoteNamesToLocalIds(remoteComponentMetadata, componentIdMapping),
 			};
 			// Pull/update already existing component + save updated component metadata
-			pullComponent(
+			await pullComponent(
 				componentType,
 				remoteComponentName,
 				componentLocalId,